+++ crossbar/controller/test/test_cli.py
@@ -142,129 +142,6 @@ class StartTests(CLITestBase):
 
         self.assertIn("Entering reactor event loop", self.stdout.getvalue())
 
-    def _test_start_run(self):
-        """
-        A basic start, that enters the reactor.
-        """
-        code_location = os.path.abspath(self.mktemp())
-        os.mkdir(code_location)
-
-        with open(self.config, "w") as f:
-            f.write("""{
-   "controller": {
-   },
-   "workers": [
-      {
-         "type": "router",
-         "options": {
-            "pythonpath": ["."]
-         },
-         "realms": [
-            {
-               "name": "realm1",
-               "roles": [
-                  {
-                     "name": "anonymous",
-                     "permissions": [
-                        {
-                           "uri": "*",
-                           "publish": true,
-                           "subscribe": true,
-                           "call": true,
-                           "register": true
-                        }
-                     ]
-                  }
-               ]
-            }
-         ],
-         "transports": [
-            {
-               "type": "web",
-               "endpoint": {
-                  "type": "tcp",
-                  "port": 8080
-               },
-               "paths": {
-            "/": {
-              "directory": ".",
-              "type": "static"
-            },
-                  "ws": {
-                     "type": "websocket"
-                  }
-               }
-            }
-         ]
-      },
-      {
-         "type": "container",
-         "options": {
-            "pythonpath": ["%s"]
-         },
-         "components": [
-            {
-               "type": "class",
-               "classname": "test.AppSession",
-               "realm": "realm1",
-               "transport": {
-                  "type": "websocket",
-                  "endpoint": {
-                     "type": "tcp",
-                     "host": "127.0.0.1",
-                     "port": 8080
-                  },
-                  "url": "ws://127.0.0.1:8080/ws"
-               }
-            }
-         ]
-      }
-   ]
-}
-            """ % ("/".join(code_location.split(os.sep),)))
-
-        with open(code_location + "/test.py", "w") as f:
-            f.write("""#!/usr/bin/env python
-from twisted.internet.defer import inlineCallbacks
-from twisted.logger import Logger
-from autobahn.twisted.wamp import ApplicationSession
-from autobahn.wamp.exception import ApplicationError
-
-class AppSession(ApplicationSession):
-
-    log = Logger()
-
-    @inlineCallbacks
-    def onJoin(self, details):
-        self.log.info("Loaded the component!")
-        yield self.publish("com.bar", "test")
-""")
-
-        reactor = SelectReactor()
-
-        def _check(lc):
-            if "Loaded the component!" in self.stdout.getvalue():
-                if reactor.running:
-                    reactor.stop()
-                lc.stop()
-
-        lc = LoopingCall(_check)
-        lc.a = (lc,)
-        lc.clock = reactor
-
-        # In case it hard-locks
-        reactor.callLater(self._subprocess_timeout, reactor.stop)
-        lc.start(0.1)
-
-        cli.run("crossbar",
-                ["start",
-                 "--cbdir={}".format(self.cbdir),
-                 "--logformat=syslogd"],
-                reactor=reactor)
-
-        self.assertIn("Entering reactor event loop", self.stdout.getvalue())
-        self.assertIn("Loaded the component!", self.stdout.getvalue())
-
     def test_configValidationFailure(self):
         """
         Running `crossbar start` with an invalid config will print a warning.
